Add bomb counts to model of minesweeper
This commit adds counts of bombs in neighbouring cells, adding a basic function to the model. Calculating cartesian product of two lists found from: https://stackoverflow.com/questions/4119730/cartesian-product-of-2-lists-in-haskell. I considered trying to turn the `Map` type into a Monad to allow for neater function use over it, when constantly getting new versions of a modified Map. However, this proved too complex, so I'm sticking with the non-Monad version for now.