Add end-game checking
Now a filter appears on top of the grid when the game is won or lost, and it is not possible to restart without reloading. Note that I tried first to send information between the clicks and the end-game grid cover using a chan, but i could not figure out how to create a fork for a UI-returning function. So, I passed the end-game cover into each tile instead.