When an empty tile is flipped, all surrounding empty tiles will also be flipped
For now, this implementation is ONLY in the model, but will be added to the front end in the next commit.
Note that an empty square is a square which is not a bomb, nor does it have any surrounding bombs. It is a standard feature that areas of empty tiles are revealed simultaneously.